api/pyproject.toml
[project]
name = "wins-smsa2p-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "minio>=7.2.15",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "secure>=1.0.1",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "plotly>=6.3.0",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "coverage>=7.6.12",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[tool.coverage.report]
omit = [
    "*/llm_engine.py"
    ]